set sub-namespace for decorator
