+++ packages/typespec-client-generator-core/test/test-host.ts
@@ -75,6 +75,7 @@ export async function createSdkTestRunner(
   const host = await createSdkTestHost(options);
   let autoUsings = [
     "Azure.ClientGenerator.Core",
+    "Azure.ClientGenerator.Core.CSharp",
     "TypeSpec.Rest",
     "TypeSpec.Http",
     "TypeSpec.Versioning",
